solving: move fenics specific blocks
