+++ dolfin_adjoint_common/blocks/__init__.py
@@ -1,6 +1,5 @@
 from .assembly import AssembleBlock
-from .solving import GenericSolveBlock, SolveVarFormBlock, SolveLinearSystemBlock
-from .projection import ProjectBlock
+from .solving import GenericSolveBlock
 from .function import FunctionAssignBlock
 from .function_assigner import FunctionAssignerBlock
 from .mesh import ALEMoveBlock, BoundaryMeshBlock
