+++ fenics_adjoint/blocks/common.py
@@ -1,8 +1,7 @@
 import dolfin_adjoint_common.blocks as blocks
 from fenics_adjoint.compat import Backend
 
-__all__ = ["AssembleBlock", "ProjectBlock", "GenericSolveBlock",
-           "SolveLinearSystemBlock", "SolveVarFormBlock",
+__all__ = ["AssembleBlock", "GenericSolveBlock",
            "FunctionAssignBlock", "DirichletBCBlock"]
 
 
@@ -10,19 +9,7 @@ class AssembleBlock(blocks.AssembleBlock, Backend):
     pass
 
 
-class ProjectBlock(blocks.ProjectBlock, Backend):
-    pass
-
-
 class GenericSolveBlock(blocks.GenericSolveBlock, Backend):
-    pass
-
-
-class SolveLinearSystemBlock(blocks.SolveLinearSystemBlock, Backend):
-    pass
-
-
-class SolveVarFormBlock(blocks.SolveVarFormBlock, Backend):
     pass
